Fixed today's date in HorizontalCalendar.tsx and ensured that content updates dynamically when other dates are selected. Also updated the related Zustand store to reflect the date changes accordingly.
